requirements.txt
backtrader>=1.9.78.123
python-binance>=1.0.19
PyYAML>=6.0
matplotlib>=3.7
pandas>=2.0
aiohttp>=3.9
requests>=2.31
